Move highlighting code to separate file (shell_highlight.cpp)
